requirements.txt
fastapi
uvicorn
scikit-learn
joblib
pydantic
langchain
langchain-core
langchain-google-genai
google-genai
langchain-chroma
chromadb
langchain-community
google-generativeai
langchain-text-splitters
pypdf
pymupdf
python-multipart
pydantic-settings